requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.3
django-cors-headers==4.6.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
iniconfig==2.0.0
packaging==24.2
phonenumbers==8.13.50
pluggy==1.5.0
PyJWT==2.10.0
sqlparse==0.5.2
tzdata==2024.2
